+++ tomcat/src/main/java/org/apache/coyote/http11/Http11Processor.java
@@ -10,7 +10,7 @@ import nextstep.jwp.exception.UncheckedServletException;
 import org.apache.coyote.Processor;
 import org.apache.coyote.http11.controller.LoginController;
 import org.apache.coyote.http11.controller.RegisterController;
-import org.apache.coyote.http11.headers.HttpHeader;
+import org.apache.coyote.http11.header.HttpHeader;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.responseline.HttpStatus;
