+++ tomcat/src/main/java/org/apache/coyote/http11/controller/LoginController.java
@@ -11,7 +11,7 @@ import java.util.UUID;
 import nextstep.jwp.db.InMemoryUserRepository;
 import nextstep.jwp.model.User;
 import org.apache.catalina.Session;
-import org.apache.coyote.http11.headers.HttpHeader;
+import org.apache.coyote.http11.header.HttpHeader;
 import org.apache.coyote.http11.request.HttpRequest;
 import org.apache.coyote.http11.response.HttpResponse;
 import org.apache.coyote.http11.responseline.HttpStatus;
